use common cache implementation
