+++ livekit-api/Cargo.toml
@@ -46,7 +46,7 @@ __signal-client-async-compatible = [
 
 
 services-tokio = ["dep:reqwest", "dep:tokio", "tokio/time"]
-services-async = ["dep:isahc"]
+services-async = ["dep:isahc", "dep:futures-timer"]
 access-token = ["dep:jsonwebtoken", "dep:hmac", "dep:signature"]
 webhooks = ["access-token", "dep:serde_json", "dep:base64"]
 
@@ -134,6 +134,8 @@ bytes = { workspace = true, optional = true }
 http = "1.1"
 reqwest = { version = "0.12", default-features = false, features = [ "json" ], optional = true }
 isahc = { version = "1.7.2", default-features = false, features = [ "json", "text-decoding" ], optional = true }
+# Runtime-agnostic timer for failover backoff on the non-tokio (isahc) backend.
+futures-timer = { version = "3", optional = true }
 
 flate2 = { version = "1", optional = true }
 scopeguard = "1.2.0"
